requirements.txt
fastapi
uvicorn
psycopg2
sqlalchemy
pydantic
python-dotenv